+++ addons/main/cfg/MSF_Factions.hpp
@@ -17,6 +17,10 @@ class CfgFactionClasses
 	{
 		displayName = "MSF - Intel System";
 	};
+	class MSF_Module_CBRN : NO_CATEGORY
+	{
+		displayName = "MSF - CBRN System";
+	};
 	class MSF_Module_Logi : NO_CATEGORY
 	{
 		displayName = "MSF - Logistics System";