+++ docs/Changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [2.14.8]
+
+- Added the groundwork for the radiation part of the CBRN system.
+- Added custom damage type to ACE Medical for radiation.
+- Added debug user addon options to enable easier toggling of MSF debug options.
+
 ## [2.14.7]
 
 - Added admin menu with common functions for logged in admins.